added to dashboard and moved around some vars
