+++ web/services/hooks/properties.tsx
@@ -1,6 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
+import { useEffect, useState } from "react";
 import { Property } from "../../lib/api/properties/properties";
+import { ok, Result } from "../../lib/result";
+import {
+  getPropertyFilters,
+  SingleFilterDef,
+} from "../lib/filters/frontendFilterDefs";
 import { getProperties } from "../lib/properties";
+import { getPropertyParams } from "../lib/propertyParams";
+import { useDebounce } from "./debounce";
 
 const useGetProperties = () => {
   const { data, isLoading, error } = useQuery({
@@ -11,11 +19,68 @@ const useGetProperties = () => {
     refetchOnWindowFocus: false,
   });
 
-  const allProperties: string[] = data?.data.map((property: Property) => {
-    return property.property;
-  });
+  const allProperties: string[] =
+    data?.data.map((property: Property) => {
+      return property.property;
+    }) ?? [];
 
-  return { properties: allProperties || [], isLoading, error };
+  const [propertyFilters, setPropertyFilters] = useState<
+    SingleFilterDef<"properties">[]
+  >(getPropertyFilters(allProperties, []));
+
+  const [propertySearch, setPropertySearch] = useState({
+    property: "",
+    search: "",
+  });
+  const debouncedPropertySearch = useDebounce(propertySearch, 500);
+
+  useEffect(() => {
+    setPropertyFilters(getPropertyFilters(allProperties, []));
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isLoading]);
+
+  useEffect(() => {
+    const { property, search } = debouncedPropertySearch;
+    async function doSearch() {
+      const values = await getPropertyParams(property, search);
+      console.log("values", values);
+      if (values.error !== null) {
+        console.error(values.error);
+        return;
+      }
+      const propertyFilters = getPropertyFilters(
+        allProperties,
+        values.data?.map((v) => ({
+          param: v.property_param,
+          key: v.property_key,
+        }))
+      );
+      console.log("propertyFilters", propertyFilters);
+      setPropertyFilters(propertyFilters);
+      console.log("searchPropertyFilters", property, search);
+    }
+
+    if (property !== "") {
+      doSearch();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [debouncedPropertySearch]);
+
+  async function searchPropertyFilters(
+    property: string,
+    search: string
+  ): Promise<Result<void, string>> {
+    setPropertySearch({ property, search });
+    return ok(undefined);
+  }
+
+  return {
+    properties: allProperties || [],
+    isLoading,
+    error,
+    propertyFilters,
+    searchPropertyFilters,
+  };
 };
 
 export { useGetProperties };
